requirements.txt
litellm
pyaudio
openwakeword
pydantic
python-dotenv
loguru 